pyproject.toml
[tool.poetry]
name = "ai-crew"
version = "0.1.0"
description = "A crew of AI agents."
authors = ["Martín Rïos martineserios@gmail.com"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
openai = "^1.33.0"
langchain = {extras = ["docarray"], version = "^0.2.5"}
docarray = "^0.40.0"
pytube = "^15.0.0"
python-dotenv = "^1.0.1"
tiktoken = "^0.7.0"
scikit-learn = "^1.5.0"
ruff = "^0.4.8"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
pypdf = "^4.2.0"
langchain-community = "^0.2.5"
langchain-pinecone = "^0.1.1"
langchain-google-genai = "^1.0.7"
pillow = "^10.3.0"
langchain-openai = "^0.1.10"
loguru = "^0.7.2"
unstructured = "^0.14.8"
markdown = "^3.6"
chromadb = "^0.5.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-asyncio = "^0.15.1"
ipykernel = "^6.29.4"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
